Pipe Grid snapping, player climbing and more assets
Created pipe grid logic allowing for drag and drop placement with the pipe snapping to the closest grid location. This grid system will be used to do the water simulation and and allow us to make gameplay changes without any coding. Also added platform and ladder with functionality and implement player climbing system.